go/runtime/bundle: Move flags to a new file
